Server/requirements.txt
fastapi
uvicorn
argon2-cffi
PyJWT